Rename some files, change calls to those files.
