+++ api/routes/uploads.js
@@ -1,9 +1,9 @@
 const express = require('express')
 const router = express.Router()
-const uploadController = require('../controllers/upload')
+const uploadsController = require('../controllers/uploads')
 const userAuthenticate = require('../middleware/user-authentication')
 
 // Get signed url
-router.get('/', userAuthenticate, uploadController.get_signed_url)
+router.get('/', userAuthenticate, uploadsController.get_signed_url)
 
 module.exports = router
